correct configs
