--- configs/_base_/schedules/schedule_160k_adamw.py
@@ -1,9 +1,0 @@
-# optimizer
-optimizer = dict(type='AdamW', lr=0.0002, weight_decay=0.0001)
-optimizer_config = dict()
-# learning policy
-lr_config = dict(policy='poly', power=0.9, min_lr=0.0, by_epoch=False)
-# runtime settings
-runner = dict(type='IterBasedRunner', max_iters=160000)
-checkpoint_config = dict(by_epoch=False, interval=4000)
-evaluation = dict(interval=4000, metric='mIoU')
